feat: rename package for project
