+++ src/test/java/io/github/popehiflo/bookhub/BookhubBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package io.github.popehiflo.bookhub_backend;
+package io.github.popehiflo.bookhub;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
